requirements.txt
flask
flask-cors
mysql-connector-python
numpy
pillow
keras-facenet
scikit-learn
opencv-python-headless
